app/requirements.txt
# FastAPI Core
fastapi[all]==0.104.1
uvicorn[standard]==0.24.0

# Authentication & Security
python-jose[cryptography]==3.3.0
cryptography==41.0.7
python-multipart==0.0.6

# Data Validation & Serialization
pydantic[email]==2.5.0
pydantic-settings==2.1.0

# HTTP Client
httpx==0.25.2

# Redis
redis==5.0.1

# Environment Management
python-dotenv==1.0.0

# Date and Time
python-dateutil==2.8.2

# Async Support
anyio==4.1.0

# Development Dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0